moniteur/app/requirements.txt
eventlet
python-engineio==3.13.2
python-socketio==4.6.1